requirements.txt
pysrt
sentence-transformers
numpy
fastapi
uvicorn
python-multipart
python-dotenv
